Complete css exercise 06
I firstly moved the .small-para selector below .para, so that it would
apply the styles in rule order.

I did the same again with .confirm, moving it below .button and again
benefitting from rule order.

I tried the same with .child for the final div elements, but it failed
to override the div.text selector's styles. To overcome this I changed
.child to .text.child and this was now specific enough to take
precedence.