Complete margin&padding exercise 01
I got as close as I could to the desired outcome. I note that the
solution seems to aim for values like 32, 48, 12, instead of round
numbers like 30, 40, 10. It's not entirely clear why.

One slight deviation was my left margin on the third div, I placed it
using a pixel value whereas the solution went with auto. Will need to
explore more to understand why auto pushed the div right over to that
right side of the available space.