Complete exercise 02 of margin and padding
A bit of a trickier one this time, I got 90% of the way there but forgot
some key options like text align.

I struggled with the margin between the content elements also, but I can
that being specific in applying it to each element is a better approach,
as in the solution.

The last hurdle was the centering of the button container at the bottom.
I was trying in vain to apply the margin 0 auto technique to the whole
container, which was effectively already centered by the margin applied
to the overall card. So I need to be more granular in future, apply text
aligns when appropriate, and then use the margin 0 auto technique on the
individual elements that require it, like the button in this case.